Changed user interaction for non-existing symbols: show the symbol in the list with error message.

+++ app/src/main/java/com/udacity/stockhawk/data/DbHelper.java
@@ -11,7 +11,7 @@ class DbHelper extends SQLiteOpenHelper {
 
 
     private static final String NAME = "StockHawk.db";
-    private static final int VERSION = 1;
+    private static final int VERSION = 2;
 
 
     DbHelper(Context context) {
@@ -27,6 +27,7 @@ class DbHelper extends SQLiteOpenHelper {
                 + Quote.COLUMN_ABSOLUTE_CHANGE + " REAL NOT NULL, "
                 + Quote.COLUMN_PERCENTAGE_CHANGE + " REAL NOT NULL, "
                 + Quote.COLUMN_HISTORY + " TEXT NOT NULL, "
+                + Quote.COLUMN_VALID + " INTEGER NOT NULL,  "
                 + "UNIQUE (" + Quote.COLUMN_SYMBOL + ") ON CONFLICT REPLACE);";
 
         db.execSQL(builder);
@@ -36,8 +37,13 @@ class DbHelper extends SQLiteOpenHelper {
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
 
-        db.execSQL(" DROP TABLE IF EXISTS " + Quote.TABLE_NAME);
+        if (newVersion == 2) {
+            db.execSQL("ALTER TABLE " + Quote.TABLE_NAME
+                    + " ADD COLUMN " + Quote.COLUMN_VALID + " INTEGER DEFAULT 0");
+        } else {
+            db.execSQL(" DROP TABLE IF EXISTS " + Quote.TABLE_NAME);
+            onCreate(db);
+        }
 
-        onCreate(db);
     }
 }
